+++ lib/ui/debug/DebugHomePanel.dart
@@ -175,7 +175,6 @@ class _DebugHomePanelState extends State<DebugHomePanel> implements Notification
                 ToggleRibbonButton(label: 'Show map levels', toggled: Storage().debugMapShowLevels!, onTap: _onMapShowLevels),
                 ToggleRibbonButton(label: 'Canvas LMS', toggled: (Storage().debugUseCanvasLms == true), onTap: _onUseCanvasLms),
                 ToggleRibbonButton(label: 'Sample Appointments', toggled: (Storage().debugUseSampleAppointments == true), onTap: _onUseSampleAppointments),
-                ToggleRibbonButton(label: 'Sample Events2', toggled: (Storage().debugUseSampleEvents2 == true), onTap: _onUseSampleEvents2),
                 ToggleRibbonButton(label: 'Mobile icard - Use Identity BB', toggled: (Storage().debugUseIdentityBb == true), onTap: _onUseIdentityBb),
                     
                 Container(color: Colors.white, child: Padding(padding: EdgeInsets.only(top: 16), child: Container(height: 1, color: Styles().colors!.surfaceAccent))),
@@ -720,12 +719,6 @@ class _DebugHomePanelState extends State<DebugHomePanel> implements Notification
   void _onUseSampleAppointments() {
     setState(() {
       Storage().debugUseSampleAppointments = (Storage().debugUseSampleAppointments != true);
-    });
-  }
-
-  void _onUseSampleEvents2() {
-    setState(() {
-      Storage().debugUseSampleEvents2 = (Storage().debugUseSampleEvents2 != true);
     });
   }