+++ lib/ui/events2/Event2DetailPanel.dart
@@ -7,7 +7,10 @@ import 'package:illinois/ext/Event2.dart';
 import 'package:illinois/ext/Explore.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Auth2.dart';
+import 'package:illinois/service/DeviceCalendar.dart';
 import 'package:illinois/ui/WebPanel.dart';
+import 'package:illinois/ui/events2/Event2AttendanceDetailPanel.dart';
+import 'package:illinois/ui/events2/Event2CreatePanel.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:illinois/utils/AppUtils.dart';
@@ -47,8 +50,9 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   // when sending the appliction to background in iOS.
   Position? _userLocation;
 
-  bool _authLoading = false; //TBD visualize
-  bool _eventLoading = false; //TBD visualize
+  bool _authLoading = false;
+  bool _registrationLoading = false;
+  bool _eventLoading = false;
 
   @override
   void initState() {
@@ -73,7 +77,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     if (name == Auth2UserPrefs.notifyFavoritesChanged) {
       setStateIfMounted(() { });
     } else if(name == Auth2.notifyLoginChanged){
-      _refreshEvent();
+      _refreshEvent(visibleProgress: true);
     }
   }
   
@@ -81,40 +85,50 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   Widget build(BuildContext context) {
     return Scaffold(body:
       Column(children: <Widget>[
-        Expanded(child:
-          CustomScrollView(slivers: <Widget>[
-            SliverToutHeaderBar(
-              flexImageUrl:  _event?.imageUrl,
-              flexImageKey: 'event-detail-default',
-              flexRightToLeftTriangleColor: Colors.white,
-            ),
-            SliverList(delegate:
-              SliverChildListDelegate([
-                Container(color: Styles().colors?.white, child:
-                  Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
-                    _badgeWidget,
-                    _categoriesWidget,
-                    Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 16), child:
-                      Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
-                        _titleWidget,
-                        _sponsorWidget,
-                        _detailsWidget,
-                      ])
-                    ),
-                  ]),
-                ),
-                Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 24), child:
-                  Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
-                  _descriptionWidget,
-                  _buttonsWidget,
-               ]))
-              ], addSemanticIndexes:false)
-            ),
-          ])
-        ),
+        Expanded(child: _content),
       ])
     );
   }
+
+  Widget get _content => _eventLoading ? _loadingContent : _eventContent;
+
+  Widget get _loadingContent {
+      return Center(child:
+          SizedBox(width: 32, height: 32, child:
+            CircularProgressIndicator(color: Styles().colors?.fillColorSecondary,)
+          )
+        );
+  }
+
+  Widget get _eventContent => CustomScrollView(slivers: <Widget>[
+    SliverToutHeaderBar(
+      flexImageUrl:  _event?.imageUrl,
+      flexImageKey: 'event-detail-default',
+      flexRightToLeftTriangleColor: Colors.white,
+    ),
+    SliverList(delegate:
+    SliverChildListDelegate([
+      Container(color: Styles().colors?.white, child:
+      Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
+        _badgeWidget,
+        _categoriesWidget,
+        Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 16), child:
+        Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
+          _titleWidget,
+          _sponsorWidget,
+          _detailsWidget,
+        ])
+        ),
+      ]),
+      ),
+      Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 24), child:
+      Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
+        _descriptionWidget,
+        _buttonsWidget,
+      ]))
+    ], addSemanticIndexes:false)
+    ),
+  ]);
 
   Widget get _badgeWidget => _isAdmin ?
   Padding(padding: EdgeInsets.symmetric(horizontal: 16), child:
@@ -189,8 +203,9 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
       ...?_speakerDetailWidget,
       ...?_priceDetailWidget,
       ...?_privacyDetailWidget,
-      ...?_adminSettingsButtonWidget,
       ...?_addToCalendarButton,
+      ...?_adminSettingsButtonWidget,
+      ...?_attendanceDetailWidget,
       ...?_contactsDetailWidget,
     ];
 
@@ -309,6 +324,19 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     return [_buildTextDetailWidget(privacyTypeTitle, "privacy"), _detailSpacerWidget];
   }
 
+  List<Widget>? get _attendanceDetailWidget {
+    if (_isAdmin || _isAttendanceTaker) {
+      return <Widget>[
+        InkWell(
+            onTap: _onTapTakeAttendance,
+            child: _buildTextDetailWidget(Localization().getStringEx('panel.event2.detail.take_attendance.title', 'Take Attendance'), 'qr', underlined: true)),
+        _detailSpacerWidget
+      ];
+    } else {
+      return null;
+    }
+  }
+
   List<Widget>? get _contactsDetailWidget{
     if(CollectionUtils.isEmpty(_event?.contacts))
       return null;
@@ -341,7 +369,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   List<Widget>? get _adminSettingsButtonWidget => _isAdmin? <Widget>[
     InkWell(onTap: _onAdminSettings, child:
-       _buildTextDetailWidget("Event Admin Settings", "settings", )),
+       _buildTextDetailWidget("Event Admin Actions", "settings", underlined: true)),
     _detailSpacerWidget
   ] : null;
 
@@ -363,8 +391,8 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     ) : Container();
   }
 
-  List<Widget>? get _urlButtonWidget => //TBD check if this is the proper url for this button
-    StringUtils.isNotEmpty(_event?.eventUrl) ? <Widget>[_buildButtonWidget(
+  List<Widget>? get _urlButtonWidget =>
+    StringUtils.isNotEmpty(_event?.eventUrl) ? <Widget>[_buildButtonWidget(//TBD remove loading from here
       title: Localization().getStringEx('panel.groups_event_detail.button.visit_website.title', 'Visit website'),
       hint: Localization().getStringEx('panel.groups_event_detail.button.visit_website.hint', ''),
       onTap: (){_onWebButton(_event?.eventUrl, analyticsName: 'Website');}
@@ -378,7 +406,8 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
         title: Localization().getStringEx('panel.event2_detail.button.unregister.title', 'Log In to Register'),
         onTap: _onLogIn,
         externalLink: false,
-        enabled: false
+        enabled: false,
+        progress: _authLoading
     )] : null;
   }
 
@@ -387,28 +416,30 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
       return null;
 
     if (_event?.registrationDetails?.type == Event2RegistrationType.internal) { //Require App registration
-        if(_event?.userRole == Event2UserRole.participant){//Already registered
+        if( _event?.userRole == Event2UserRole.participant){//Already registered
           return <Widget>[_buildButtonWidget(
               title: Localization().getStringEx('panel.event2_detail.button.unregister.title', 'Unregister me'),
               onTap: _onUnregister,
               externalLink: false,
-              enabled: false
+              enabled: false,
+              progress: _registrationLoading
           )];
         } else if (_event?.userRole == null){//Not registered yet
           return <Widget>[_buildButtonWidget(
               title: Localization().getStringEx('panel.event2_detail.button.register.title', 'Register me'),
               onTap: _onRegister,
-              externalLink: false
+              externalLink: false,
+              progress: _registrationLoading,
           )];
         }
     } else if(StringUtils.isNotEmpty(_event?.registrationDetails?.externalLink)){// else external registration
-      if(_event?.userRole == null){ //TBD check if this is correct check or we don't know if the user is registered externally
+      // if(_event?.userRole == null){ //TBD check if this is correct check or we don't know if the user is registered externally
         return <Widget>[_buildButtonWidget(
             title: Localization().getStringEx('panel.event2_detail.button.register.title', 'Register me'),
             onTap: _onExternalRegistration,
             externalLink: true
         )];
-      }
+      // }
     }
 
     return null; //not required
@@ -431,10 +462,10 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   Widget _buildTextDetailWidget(String text, String iconKey, {
     EdgeInsetsGeometry contentPadding = const EdgeInsets.only(top: 4),
     EdgeInsetsGeometry iconPadding = const EdgeInsets.only(right: 6),
-    bool iconVisible = true, bool underlined = false
+    bool iconVisible = true, bool underlined = false, int maxLines = 1,
   }) =>
     _buildDetailWidget(
-      Text(text, maxLines: 1, style: underlined? Styles().textStyles?.getTextStyle('widget.info.medium.underline') : Styles().textStyles?.getTextStyle('widget.info.medium'),),
+      Text(text, style: underlined? Styles().textStyles?.getTextStyle('widget.info.medium.underline') : Styles().textStyles?.getTextStyle('widget.info.medium'), maxLines: maxLines, overflow: TextOverflow.ellipsis,),
       iconKey,
       contentPadding: contentPadding,
       iconPadding: iconPadding,
@@ -467,7 +498,8 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
     String? hint,
     bool enabled = true,
     bool externalLink = true,
-    void Function()? onTap
+    bool progress = false,
+    void Function()? onTap,
   }) => StringUtils.isNotEmpty(title) ?
     Padding(padding: EdgeInsets.only(bottom: 6), child:
       Row(children:<Widget>[
@@ -481,10 +513,11 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
               rightIcon:externalLink? Styles().images?.getImage(enabled ? 'external-link' : 'external-link-dark' ) : null,
               padding: EdgeInsets.symmetric(vertical: 8, horizontal: 8),
               onTap: onTap ?? (){},
+            progress: progress,
             contentWeight: 0.5,
-
           ),
-        ),],)
+        ),
+        ] )
       ) : Container();
   
   Widget _buildSettingButton({String? title, VoidCallback? onTap}) =>  StringUtils.isNotEmpty(title) ?
@@ -506,6 +539,10 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   void _onOnline() {
     Analytics().logSelect(target: "Online Url: ${_event?.name}");
+    String? url = _event?.onlineDetails?.url;
+    if(StringUtils.isNotEmpty(url)){
+      _onLaunchUrl(url);
+    }
   }
 
   void _onFavorite() {
@@ -516,7 +553,7 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   void _onLaunchUrl(String? url, {BuildContext? context}) {
     if (StringUtils.isNotEmpty(url)) {
       if (UrlUtils.launchInternal(url)) {
-        Navigator.push(context!, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
+        Navigator.push(context ?? this.context, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
       } else {
         Uri? uri = Uri.tryParse(url!);
         if (uri != null) {
@@ -537,28 +574,32 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   void _onRegister(){
     if(_event?.id != null){
-      Events2().registerToEvent(_event!.id!).then((value){
-          if(value != null){ //success
-            String? errorMessage = JsonUtils.stringValue(value);
-            if(StringUtils.isNotEmpty(errorMessage)){
-              Log.e(errorMessage);
-            } else {
-              _refreshEvent();
-            }
+      setStateIfMounted(() { _registrationLoading = true;});
+      Events2().registerToEvent(_event!.id!).then((errorMessage){
+        setStateIfMounted(() { _registrationLoading = false;});
+        if(errorMessage == null){
+          _refreshEvent();
+        } else {
+          String? message = JsonUtils.stringValue(errorMessage);
+          if(StringUtils.isNotEmpty(message)){
+            Log.e(message);
           }
+        }
       });
     }
   }
 
   void _onUnregister(){
     if(_event?.id != null){
-      Events2().unregisterFromEvent(_event!.id!).then((value){
-        if(value != null){ //success
-          String? errorMessage = JsonUtils.stringValue(value);
-          if(StringUtils.isNotEmpty(errorMessage)){
-            Log.e(errorMessage);
-          } else {
-            _refreshEvent();
+      setStateIfMounted(() { _registrationLoading = true;});
+      Events2().unregisterFromEvent(_event!.id!).then((errorMessage){
+        setStateIfMounted(() { _registrationLoading = false;});
+        if(errorMessage == null){
+          _refreshEvent();
+        } else { //fail
+          String? message = JsonUtils.stringValue(errorMessage);
+          if(StringUtils.isNotEmpty(message)){
+            Log.e(message);
           }
         }
       });
@@ -582,20 +623,23 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
         }
       );
     }
-  }
-
+  }  
   void _onAddToCalendar(){
-    //TBD
+      DeviceCalendar().addToCalendar(_event);
   }
 
   void _onContactEmail(String? email){
-    //TBD
-    AppToast.show("TBD send email to $email");
+    if(StringUtils.isNotEmpty(email)) {
+      String link = "mailto:$email";
+      _onLaunchUrl(link);
+    }
   }
 
   void _onContactPhone(String? phone){
-    //TBD
-    AppToast.show("TBD call to $phone");
+    if(StringUtils.isNotEmpty(phone)) {
+      String link = "tel:$phone";
+      _onLaunchUrl(link);
+    }
   }
 
   void _onAdminSettings(){
@@ -612,7 +656,8 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   }
 
   void _onSettingEditEvent(){
-    //TBD
+    Analytics().logSelect(target: "Edit event");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => Event2CreatePanel(event: _event)));
   }
 
   void _onSettingPromote(){
@@ -632,14 +677,33 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   }
 
   void _onSettingDeleteEvent(){
-    //TBD
+    Analytics().logSelect(target: 'Delete Event');
+    if(_eventId != null) {
+      _eventLoading = true;
+      Events2().deleteEvent(_eventId!).then((errorMessage) {
+        if(errorMessage == null){
+          setStateIfMounted(() {_eventLoading = false; });
+          Navigator.of(context).pop();
+        } else {
+          String? message = JsonUtils.stringValue(errorMessage);
+          if(StringUtils.isNotEmpty(message)){
+            Log.e(message);
+          }
+        }
+      });
+    }
+  }
+
+  void _onTapTakeAttendance() {
+    Analytics().logSelect(target: 'Take Attendance');
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => Event2AttendanceDetailPanel(event: _event)));
   }
 
   //loading
   void _initEvent() async {
     _event = widget.event;
 
-    if(_event == null && StringUtils.isEmpty(widget.eventId!)) {
+    if(_event == null && StringUtils.isNotEmpty(widget.eventId!)) {
       _eventLoading = true;
       _loadEvent().then((event) {
         _eventLoading = false;
@@ -664,14 +728,17 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
       // return Do we allow it?
     }
 
-    String? eventId = _event?.id ?? widget.eventId;
-    if(eventId != null) {
-      return Events2().loadEvent(eventId);
+    if(_eventId != null) {
+      return Events2().loadEvent(_eventId!);
     }
     return null;
   }
 
   //Event getters
   bool get _isAdmin =>  _event?.userRole == Event2UserRole.admin;
+
+  bool get _isAttendanceTaker =>  _event?.userRole == Event2UserRole.attendanceTaker;
+
+  String? get _eventId => _event?.id ?? widget.eventId;
 
 }